requirements.txt
Flask==3.0.3
numpy==1.26.4
matplotlib==3.8.4
pillow==10.3.0
requests==2.31.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.18.1+cpu
